requirements.txt
pyg==2.0.3
cudatoolkit==11.3.1
cudnn==8.2.1
python==3.9.7
pytorch==1.10.1
mkl==2024.0.0
pandas
tqdm
pooch
joblib
scikit-learn==0.24.2
crest>=2.12
numpy>=1.19.4
nwchem>=7.0.2
openbabel>=3.0.0
openmpi
pandas>=1.1.4
python==3.9
rdkit>=2023.09.1
snakemake>=6.3.0
statsmodels>=0.11.1
xtb>=6.5.1
git
alfabet
tensorflow>=2.0.0
keras<=2.15.0
